pyproject.toml
[tool.poetry]
name = "discord-undoc"
version = "0.1.0"
description = "Unofficial Discsord Documentation - Documenting the undocumented side of the Discord API"
authors = ["arHSM <hanseungmin.ar@gmail.com>"]
license = "Apache-2.0"

[tool.poetry.dependencies]
python = "^3.10"
PyYAML = "^6.0"

[tool.poetry.dev-dependencies]
black = "^22.1.0"
isort = "^5.10.1"
pyright = "^0.0.13"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
